Rename Settings and refactor Profile
