+++ src/pages/Profile.js
@@ -1,16 +1,12 @@
 import React from "react"
-import { Link } from "react-router-dom"
 
 import { default as Layout } from "../templates/LayoutSingle"
 
+import SegmentProfile from "../large/SegmentProfile"
+
 const Profile = props => (
   <Layout>
-    <h2>Profile of {props.match.params.username}</h2>
-    <div>
-      <Link to="/setting">
-        <button>Go to setting</button>
-      </Link>
-    </div>
+    <SegmentProfile />
   </Layout>
 )
