+++ src/styles/index.css
@@ -28,8 +28,7 @@ a {
   color: black;
 }
 
-a:visited,
-a:hover {
+a:visited, a:hover {
   text-decoration: none;
   color: black;
 }
@@ -252,7 +251,6 @@ a:hover {
   padding-left: 12px;
   padding-right: 12px;
   width: 100%;
-  max-width: 600px;
 }
 
 /* -------------------------------------------------------------------------- */
